Задание 5 и 7 (доработка. НЕ окончательный вариант!)

+++ src/main/java/Calculator/Calculator.java
@@ -1,5 +1,5 @@
 package Calculator;
 
 public interface Calculator {
-    int calc(int num1, int num2);
+    double calc(double num1, double num2);
 }
