+++ src/main/java/Calculator/Multiplication.java
@@ -5,7 +5,7 @@ public class Multiplication implements Calculator {
     }
 
     @Override
-    public int calc(int num1, int num2) {
+    public double calc(double num1, double num2) {
         return num1 * num2;
     }
 }
